Refactor benchmark (#163)
* refactor benchmark folders structure
* refactor ApiMock usage: create BaseApiMock
