+++ test/integration/ma_crossover.js
@@ -3,20 +3,12 @@
 'use strict'
 
 const { delay } = require('bluebird')
-const { Candle } = require('bfx-api-node-models')
 const { expect } = require('chai')
 
-const { spawnHost, connectHost, destroyHost } = require('../../benchmarks/fixtures/host')
-const ApiMock = require('../util/mimic/bitfinex_api_mock')
-const { candles: candlesDataProvider } = require('../util/mimic/data-providers')
-const Randomizer = require('../util/randomizer')
-const authHandler = require('../util/mimic/handlers/auth')
-const pingHandler = require('../util/mimic/handlers/ping')
-const subscribeToChannelHandler = require('../util/mimic/handlers/subscribe_to_channel')
-const newOrderHandler = require('../util/mimic/handlers/new_order')
-const unsubscribeHandler = require('../util/mimic/handlers/unsubscribe')
+const { spawnHost, connectHost, destroyHost } = require('../util/host')
 const { NEW_ORDER } = require('../util/mimic/signal_types')
 const ApiSpy = require('../util/mimic/testing/api_spy')
+const BaseApiMock = require('../util/base_api_mock')
 
 function performOrder (host) {
   return host.startAO('bfx-ma_crossover', {
@@ -40,31 +32,11 @@ describe('ema integration test', () => {
   let host, apiMock, spyServer
   const apiKey = 'api key'
   const apiSecret = 'api secret'
-  const baseCandle = new Candle({
-    mts: 0,
-    open: 12_000,
-    close: 15_000,
-    high: 16_000,
-    low: 9_000,
-    volume: 3
-  })
-  const seed = 59
-  const randomizer = new Randomizer(seed)
-  const dataProviders = {
-    candles: candlesDataProvider(randomizer.fork(), baseCandle.serialize())
-  }
 
   before(async () => {
-    apiMock = new ApiMock({
-      session: {
-        eventHandlers: {
-          auth: authHandler(event => event.apiKey === apiKey),
-          ping: pingHandler,
-          subscribe: subscribeToChannelHandler(dataProviders),
-          unsubscribe: unsubscribeHandler,
-          [NEW_ORDER]: newOrderHandler
-        }
-      }
+    apiMock = new BaseApiMock({
+      apiKey,
+      dataProviders: BaseApiMock.getDataProviders()
     })
 
     spyServer = new ApiSpy(apiMock)
